Refactor Cache
Cache is back to using the read/write buffer that does its bus access for it.

It has a "ChangeDetector" component that can tell when input wires change, so it doesn't flood the write buffer with loads.

It stalls properly and can wait for the buffer to flush.

Next steps:
* write back dirty lines proactively.
* Lookup read requests in the write buffer on cache misses.